bootstrap_tf class

A list object returned by bootstrap_trendfilter(), with the following elements:

tf_standard_errors The standard errors of the optimized trend filtering point estimator.

bootstrap_lower_band Vector of lower bounds for the pointwise variability bands, evaluated on x_eval.

bootstrap_upper_band Vector of upper bounds for the pointwise variability bands, evaluated on x_eval.

bootstrap_algorithm A string specifying which variation of the bootstrap was used to obtain the variability bands.

level The level of the pointwise variability bands.

B The number of bootstrap samples used to estimate the pointwise variability bands.

tf_bootstrap_ensemble If return_ensemble = TRUE, the full trend filtering bootstrap ensemble as an \(n \times B\) matrix, less any columns from post-hoc pruning (if prune = TRUE). Else, this will return NULL.

edf_boots An integer vector of the estimated number of effective degrees of freedom of each trend filtering bootstrap estimate. These should all be relatively close to edf_min (below).

prune Logical. If TRUE, then the trend filtering bootstrap ensemble is examined for rare instances in which the optimization has stopped at zero knots (likely erroneously), and removes them from the ensemble.

n_pruned The number of poorly-converged bootstrap trend filtering estimates pruned from the ensemble.

n_iter_boots Vector of the number of iterations needed for the ADMM algorithm to converge within the given tolerance, for each bootstrap trend filtering estimate.

...: Named elements inherited from obj --- an object either of class sure_tf or cv_tf. See the relevant function documentation for details.

See also