Construct pointwise variability bands via a tailored bootstrap algorithm

Generate a bootstrap ensemble of trend filtering estimates in order to quantify the uncertainty in the optimized estimate. One of three possible bootstrap algorithms should be chosen according to the criteria in the details section below. See Politsch et al. (2020a) for the technical details of each bootstrap algorithm. Pointwise variability bands are then obtained by passing the bootstrap_trendfilter() output and a desired level (e.g. level = 0.95) to vbands().

bootstrap_trendfilter(
  obj,
  bootstrap_algorithm = c("nonparametric", "parametric", "wild"),
  B = 100L,
  mc_cores = parallel::detectCores() - 4
)

Arguments

obj
An object of class 'pred_tf'.

bootstrap_algorithm
A string specifying which variation of the bootstrap to use. One of c("nonparametric", "parametric", "wild"). See details section below for when each option is appropriate.

B
The number of bootstrap samples used to estimate the pointwise variability bands. Defaults to B = 100L.

mc_cores
Multi-core computing using the parallel package: The number of cores to utilize. Defaults to the number of cores detected, minus 4.

Value

An object of class 'bootstrap_tf'. This is a list with the following elements

x_eval

Input grid that each bootstrap trend filtering estimate was evaluated on.

ensemble

The full trend filtering bootstrap ensemble as an \(n \times B\) matrix.

edf_boots

Vector of the estimated number of effective degrees of freedom of each trend filtering bootstrap estimate. These should all be relatively close to obj$edf_opt.

n_iter_boots

Vector of the number of iterations taken by the ADMM algorithm before reaching a stopping criterion, for each bootstrap estimate.

lambda_boots

Vector of the hyperparameter values used for each bootstrap fit. In general, these are not all equal because our bootstrap implementation instead seeks to hold the number of effective degrees of freedom constant across all bootstrap estimates.

bootstrap_algorithm

A string specifying which variation of the bootstrap was used to generate the ensemble.

Details

Our recommendations for when to use each of the possible settings for the bootstrap_algorithm argument are shown in the table below. See Politsch et al. (2020a) for more details.

Scenario	Uncertainty quantification

x is unevenly sampled	bootstrap_algorithm = "nonparametric"

x is evenly sampled and measurement variances for y are available	bootstrap_algorithm = "parametric"

x is evenly sampled and measurement variances for y are not available	bootstrap_algorithm = "wild"

For our purposes, an evenly sampled data set with some discarded pixels (either sporadically or in large consecutive chunks) is still considered to be evenly sampled. When the inputs are evenly sampled on a transformed scale, we recommend transforming to that scale and carrying out the full trend filtering analysis on that scale. See Example 2 below for a case when the inputs are evenly sampled on the log10(x) scale.

References

Companion references

Politsch et al. (2020a). Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy. MNRAS, 492(3), p. 4005-4018. [arXiv].

Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical signals with varying degrees of smoothness. MNRAS, 492(3), p. 4019-4032. [arXiv].

The Bootstrap and variations

Efron and Tibshirani (1986). Bootstrap Methods for Standard Errors, Confidence Intervals, and Other Measures of Statistical Accuracy. Statistical Science, 1(1), p. 54-75.

Mammen (1993). Bootstrap and Wild Bootstrap for High Dimensional Linear Models. The Annals of Statistics, 21(1), p. 255-285.

Efron (1979). Bootstrap Methods: Another Look at the Jackknife. The Annals of Statistics, 7(1), p. 1-26.

See also

Examples

# Example 1: Phase-folded light curve of an eclipsing binary star system
#
# The apparent brightness over time of a star system that has two suns
# that regularly eclipse one another from our vantage point on Earth. Here,
# the time series is stacked according to the orbital period of the binary
# system, with the primary eclipse occuring at `phase = 0` and the input
# domain ranging from -0.5 to 0.5.

data(eclipsing_binary)
head(EB)

#> # A tibble: 6 × 3
#>    phase  flux std_err
#>    <dbl> <dbl>   <dbl>
#> 1 -0.499 0.938  0.0102
#> 2 -0.498 0.930  0.0102
#> 3 -0.496 0.944  0.0102
#> 4 -0.494 0.963  0.0102
#> 5 -0.494 0.934  0.0102
#> 6 -0.494 0.949  0.0102

cv_tf <- cv_trendfilter(
  x = EB$phase,
  y = EB$flux,
  weights = 1 / EB$std_err^2,
  optimization_params = list(
    max_iter = 1e4,
    obj_tol = 1e-6,
    thinning = TRUE
  )
)

pred_tf <- predict(
  cv_tf,
  loss_func = "MAE",
  lambda_choice = "lambda_1se",
  nx_eval = 1500L
)

boot_tf <- bootstrap_trendfilter(pred_tf, "nonparametric")
bands <- vbands(boot_tf)

# Example 2: The "Lyman-alpha forest" in the spectrum of a distant quasar

data(quasar_spectrum)
head(spec)

#> # A tibble: 6 × 3
#>   log10_wavelength   flux weights
#>              <dbl>  <dbl>   <dbl>
#> 1             3.55  0.424  0.0417
#> 2             3.55 -2.11   0.125 
#> 3             3.55 -3.78   0.128 
#> 4             3.55  3.60   0.121 
#> 5             3.55 -4.89   0.134 
#> 6             3.55  1.76   0.132 

sure_tf <- sure_trendfilter(spec$log10_wavelength, spec$flux, spec$weights)
pred_tf <- predict(sure_tf)

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Error in if (any(x_eval < min(obj$tf_model$x) || x_eval > max(obj$tf_model$x))) {    stop("x_eval should all be in range(x).")}: missing value where TRUE/FALSE needed

boot_tf <- bootstrap_trendfilter(pred_tf, "parametric")
bands <- vbands(boot_tf)