cv_tf class

A list object returned by cv_trendfilter(), with the following elements:

x_eval: Input grid used to evaluate the optimized trend filtering estimate on.

tf_estimate: Optimized trend filtering estimate, evaluated at x_eval.

validation_method: paste0(V,"-fold CV")

validation_functional: Type of error that validation was performed on. Either one of c("WMAE","WMSE","MAE","MSE") or a custom function passed by the user.

V: The number of folds the data are split into for the V-fold cross validation.

lambdas: Vector of hyperparameter values evaluated in the grid search (always returned in descending order).

edfs: Vector of effective degrees of freedom for all trend filtering estimators fit during validation.

generalization_errors: Vector of cross validation estimates of the trend filtering generalization error, for each hyperparameter value (ordered corresponding to the descending-ordered lambdas vector).

se_errors: The standard errors of the cross validation errors. These are particularly useful for implementing the ``1-standard-error rule''.

lambda_min: Hyperparameter value that minimizes the cross validation generalization error curve.

`lambda_1se`: Largest hyperparameter value that is within one standard error of the minimum hyperparameter's cross validation error.

`lambda_choice`: One of `c("lambda_min", "lambda_1se")`. The choice of hyperparameter that is used for the returned trend filtering estimate evaluation `tf_estimate`.

`i_min`: Index of `lambdas` that minimizes the cross validation error.

`i_1se`: Index of `lambdas` that gives the largest hyperparameter value that has a cross validation error within 1 standard error of the minimum of the cross validation error curves.

`edf_min`: Effective degrees of freedom of the optimized trend filtering estimator.

`edf_1se`: Effective degrees of freedom of the 1-stand-error rule trend filtering estimator.

`n_iter`: The number of iterations needed for the ADMM algorithm to converge within the given tolerance, for each hyperparameter value. If many of these are exactly equal to `max_iter`, then their solutions have not converged with the tolerance specified by `obj_tol`. In which case, it is often prudent to increase `max_iter`.

`x`: Vector of observed inputs.

`y:` Vector of observed outputs.

`weights`: Weights for the observed outputs, defined as the reciprocal variance of the additive noise that contaminates the signal.

`fitted_values`: Optimized trend filtering estimate, evaluated at the observed inputs `x`.

`residuals`: `residuals = y - fitted_values`

`k`: Degree of the trend filtering estimator.

`admm_params`: List of parameter settings for the trend filtering ADMM algorithm, constructed by passing the `optimization_params` list to glmgen::trendfilter.control.list().

`thinning`: Logical. If `TRUE`, then the data are preprocessed so that a smaller, better conditioned data set is used for fitting.

`x_scale`, `y_scale`, `data_scaled`: For internal use.

See also