Optimize the trend filtering hyperparameter by V-fold cross validation

For every candidate hyperparameter value, estimate the corresponding trend filtering model's out-of-sample error by V-fold cross validation. Four commonly-used types of error --- MSE, WMSE, MAE, WMAE --- are all evaluated and returned, and the user has the option to pass additional error functionals to be evaluated. See the details section for when you should use cv_trendfilter() versus sure_trendfilter().

cv_trendfilter(
  x,
  y,
  weights,
  k = 2L,
  nlambdas = 250L,
  V = 10L,
  mc_cores = parallel::detectCores() - 4,
  custom_error_funcs,
  optimization_params
)

Arguments

x
Vector of observed values for the input variable.

y
Vector of observed values for the output variable.

weights
Weights for the observed outputs, defined as the reciprocal variance of the additive noise that contaminates the signal in y. When the noise is expected to have equal variance for all observations, weights can be passed as a scalar. Otherwise, weights must be a vector with the same length as x and y.

k
Degree of the polynomials that make up the piecewise-polynomial trend filtering estimate. Defaults to k = 2 (i.e. a piecewise quadratic estimate). Must be one of k = 0,1,2. Higher order polynomials are disallowed since their smoothness is indistinguishable from k = 2 and their use can lead to instability in the convex optimization.

nlambdas
Number of hyperparameter values to test during validation. Defaults to nlambdas = 250. The hyperparameter grid is dynamically constructed to span the full model space lying between a single polynomial solution (i.e. a power law) and an interpolating solution, with nlambdas controlling the granularity of the hyperparameter grid.

V
Number of folds that the data are partitioned into for the V-fold cross validation. Defaults to V = 10.

mc_cores
Multi-core computing using the parallel R package: The number of cores to utilize. Defaults to the number of cores detected, minus four.

custom_error_funcs
(Optional) A named list of one or more functions, with each defining an error functional to evaluate on held-out folds during cross validation. Mean-squared error (MSE) and mean absolute error (MAE) are both evaluated automatically, as well as weighted versions of each (with reciprocal variances as weights) --- WMSE and WMAE. Therefore, the user does not need to pass anything to custom_error_funcs unless they want to define a validation error metric other than MSE, MAE, WMSE, and WMAE.

In such a case, each function in the named list passed to custom_error_funcs should take three vector arguments --- y, tf_estimate, and weights --- and return a single scalar value for the validation error. For example, if I also wanted to define a validation error curve from the weighted median of the absolute errors, I could pass the following list to custom_error_funcs:

list(MedAE = function(tf_estimate, y, weights) {
               matrixStats::weightedMedian(abs(tf_estimate - y), weights)
             }
    )

optimization_params
(Optional) A named list of parameter values to be passed to the trend filtering ADMM algorithm of Ramdas and Tibshirani (2016), which is implemented in the glmgen R package. See the glmgen::trendfilter.control.list() documentation for a full list of the algorithm's parameters. The default parameter choices will almost always suffice, but when adjustments do need to be made, one can do so without any technical understanding of the ADMM algorithm. In particular, the four parameter descriptions below should serve as sufficient working knowledge.

obj_tol

A stopping threshold for the ADMM algorithm. If the relative change in the algorithm's cost functional between two consecutive steps is less than obj_tol, the algorithm terminates. The algorithm's termination can also result from it reaching the maximum tolerable iterations set by the max_iter parameter (see below). The obj_tol parameter defaults to obj_tol = 1e-10. The cost_functional vector, returned within the sure_trendfilter() output, gives the relative change in the trend filtering cost functional over the algorithm's final iteration, for every candidate hyperparameter value.

max_iter

Maximum number of ADMM iterations that we will tolerate. Defaults to max_iter = length(y). The actual number of iterations performed by the algorithm, for every candidate hyperparameter value, is returned in the n_iter vector, within the sure_trendfilter() output. If any of the elements of n_iter are equal to max_iter, the tolerance defined by obj_tol has not been attained and max_iter may need to be increased.

thinning

Logical. If thinning = TRUE, then the data are preprocessed so that a smaller data set is used to fit the trend filtering estimate, which will ease the ADMM algorithm's convergence. This can be very useful when a signal is so well-sampled that very little additional information / predictive accuracy is gained by fitting the trend filtering estimate on the full data set, compared to some subset of it. See the cv_trendfilter() examples for a case study of this nature. When nothing is passed to thinning, the algorithm will automatically detect whether thinning should be applied. This preprocessing procedure is controlled by the x_tol parameter below.

x_tol

Controls the automatic detection of when thinning should be applied to the data. If we make bins of size x_tol and find at least two elements of x that fall into the same bin, then the data is thinned.

Value

An object of class 'cv_tf'. This is a list with the following elements:

lambdas

Vector of candidate hyperparameter values (always returned in descending order).

edfs

Number of effective degrees of freedom in the trend filtering estimator, for every hyperparameter value in lambdas.

validation_errors

A named list of vectors, with each representing the cross validation error curve for some definition of error. The first 4 vectors of the list correspond to WMAE, WMSE, MAE, MSE. If any custom error functionals were passed to custom_error_funcs, their cross validation curves will follow the first 4.

se_validation_errors

Named list of estimated standard errors for each of the cross validation error curves in validation_errors.

lambda_min

A named vector with length equal to length(validation_errors), containing the hyperparameter value that minimizes the cross validation error curve, for every type validation error.

lambda_1se

A named vector with length equal to length(validation_errors), containing the "1-standard-error rule" hyperparameter, for every type validation error. The "1-standard-error rule" hyparameter is the largest hyperparameter value (corresponding to the smoothest trend filtering estimate) that has a cross validation error within one standard error of the minimum cross validation error. It serves as an Occam's razor-esque heuristic. That is, given two models with approximately equal performance, it may be wise to opt for the simpler model, i.e. the model with fewer effective degrees of freedom.

edf_min

A named vector with length equal to length(validation_errors), containing the number of effective degrees of freedom in the trend filtering estimator that minimizes the cross validation error curve, for every type of validation error.

edf_1se

A named vector with length equal to length(validation_errors), containing the number of effective degrees of freedom in the "1-standard-error rule" trend filtering estimator, for every type of validation error.

i_min

A named vector with length equal to length(validation_errors), containing the index of lambdas that minimizes the CV error curve, for every type of validation error.

i_1se

A named vector with length equal to length(validation_errors), containing the index of lambdas that gives the "1-standard-error rule" hyperparameter value, for every type of validation error.

validation_error_funcs

A named list of functions that define the types of error that were evaluated during cross validation.

cost_functional

The relative change in the cost functional over the ADMM algorithm's final iteration, for every candidate hyperparameter in lambdas.

n_iter

Total number of iterations taken by the ADMM algorithm, for every candidate hyperparameter in lambdas. If an element of n_iter is exactly equal to max_iter, then the ADMM algorithm stopped before reaching the tolerance set by obj_tol. In these cases, you may need to increase max_iter to ensure the trend filtering solution has converged to satisfactory precision.

V

The number of folds the data were split into for cross validation.

model_fit

A list of objects that is used internally by other functions that operate on the cv_trendfilter() output.

Details

Our recommendations for when to use cv_trendfilter() versus sure_trendfilter() are shown in the table below.

Scenario	Hyperparameter optimization

x is unevenly sampled	cv_trendfilter()

x is evenly sampled and reciprocal variances are not available	cv_trendfilter()

x is evenly sampled and reciprocal variances are available	sure_trendfilter()

For our purposes, an evenly sampled data set with some discarded pixels (either sporadically or in large consecutive chunks) is still considered to be evenly sampled. When the inputs are evenly sampled on a transformed scale, we recommend transforming to that scale and carrying out the full trend filtering analysis on that scale. See the sure_trendfilter() examples for a case when the inputs are evenly sampled on the log10(x) scale.

The validation error functionals that we automatically compute CV curves for are formally defined below.

\[MSE(\lambda) = \frac{1}{n} \sum_{i=1}^{n} |Y_i - \hat{f}(x_i; \lambda)|^2\] \[MAE(\lambda) = \frac{1}{n} \sum_{i=1}^{n} |Y_i - \hat{f}(x_i; \lambda)|\] \[WMSE(\lambda) = \sum_{i=1}^{n} |Y_i - \hat{f}(x_i; \lambda)|^2\frac{w_i}{\sum_jw_j}\] \[WMAE(\lambda) = \sum_{i=1}^{n} |Y_i - \hat{f}(x_i; \lambda)|\frac{\sqrt{w_i}}{\sum_j\sqrt{w_j}}\] where \(w_i:=\) weights[i].

If constant weights are passed, or if nothing is passed, then the weighted and unweighted counterparts are equivalent.

Briefly stated, weighting the validation error metric (with reciprocal variances) helps prevent the error from being dominated by the model's (in)accuracy on a (potentially very small) subset of the data that have high variance; and absolute error is less sensitive to outliers than squared error.

References

Companion references

Politsch et al. (2020a). Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy. MNRAS, 492(3), p. 4005-4018.

Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical signals with varying degrees of smoothness. MNRAS, 492(3), p. 4019-4032.

Cross validation

Hastie, Tibshirani, and Friedman (2009). The Elements of Statistical Learning: Data Mining, Inference, and Prediction. 2nd edition. Springer Series in Statistics. (See Sections 7.10 and 7.12)

See also

Examples

data(eclipsing_binary)
head(EB)

#> # A tibble: 6 × 3
#>    phase  flux std_err
#>    <dbl> <dbl>   <dbl>
#> 1 -0.499 0.938  0.0102
#> 2 -0.498 0.930  0.0102
#> 3 -0.496 0.944  0.0102
#> 4 -0.494 0.963  0.0102
#> 5 -0.494 0.934  0.0102
#> 6 -0.494 0.949  0.0102

cv_tf <- cv_trendfilter(
  x = EB$phase,
  y = EB$flux,
  weights = 1 / EB$std_err^2,
  optimization_params = list(
    max_iter = 1e4,
    obj_tol = 1e-6,
    thinning = TRUE
  )
)