Predict method for trend filtering fits

Predict values based on a sure_tf or cv_tf object.

# S3 method for sure_tf
predict(obj, lambda, x_eval, nx_eval)

Arguments

obj
An object of class sure_tf or cv_tf.

x_eval
(Optional) Overrides nx_eval if passed. A grid of inputs to evaluate the optimized trend filtering estimate on.

nx_eval
Integer. The length of the input grid that the optimized trend filtering estimate is evaluated on; i.e. if nothing is passed to x_eval, then it is defined as x_eval = seq(min(x), max(x), length = nx_eval).

Value

A list with the following elements:

x_eval

Input grid used to evaluate the optimized trend filtering estimate on.

tf_estimate

Optimized trend filtering estimate, evaluated at x_eval.

fitted_values

Optimized trend filtering estimate, evaluated at the observed inputs x.

residuals

residuals = y - fitted_values