Evaluate an optimized trend filtering estimate

Evaluate an optimized trend filtering estimate on a grid of input values. The predict function can be called on an object of class 'cv_tf' or 'sure_tf', and the arguments validation_error_metric (for 'cv_tf' objects only) and lambda_choice should be used to specify the desired method for optimizing the trend filtering hyperparameter.

# S3 method for cv_tf
predict(
  obj,
  validation_error_metric = "WMAE",
  lambda_choice = c("lambda_min", "lambda_1se"),
  x_eval,
  nx_eval
)

# S3 method for sure_tf
predict(obj, lambda_choice = c("lambda_min", "lambda_1se"), x_eval, nx_eval)

Arguments

obj
An object of class 'cv_tf' or 'sure_tf'.

validation_error_metric
(For class 'cv_tf' only) A string or index specifying which cross validation error curve stored within the cv_tf object will be used to choose the trend filtering hyperparameter. The first 4 options are "WMAE", "WMSE", "MAE", and "MSE". Therefore, for example, one could select the CV curve based on mean absolute deviations error by passing either validation_error_metric = "MAE" or validation_error_metric = 3. If the user passed custom error functions to cv_trendfilter(), then these will be available as additional options following the first 4, and their string designations can be seen by running names(obj$validation_error_funcs).

lambda_choice
One of c("lambda_min", "lambda_1se"). The choice of hyperparameter that is used for optimized trend filtering estimate. Defaults to lambda_choice = "lambda_min".

"lambda_min": The hyperparameter value that minimizes the cross validation error curve.

"lambda_1se": The largest hyperparameter value with a cross validation error within 1 standard error of the minimum cross validation error. This choice therefore favors simpler (i.e. smoother) trend filtering estimates.

x_eval
A grid of inputs to evaluate the optimized trend filtering estimate on. Defaults to the observed values x when nothing is passed.

nx_eval
Integer. If passed, overrides x_eval with x_eval = seq(min(x), max(x), length = nx_eval).

Value

An object of class 'pred_tf'. This is a list with the following elements:

x_eval

tf_estimate

lambdas

edfs

validation_error_metric

validation_errors

se_validation_errors

lambda_opt

edf_opt

i_opt

tf_model

Details

The motivation for using lambda_choice = "lambda_1se" is essentially Occam's razor: the two models yield results that are quantitatively very close, so we favor the simpler model. See Section 7.10 of Hastie, Tibshirani, and Friedman (2009) for more details on the "one-standard-error rule".

See also

Examples

data(eclipsing_binary)
head(EB)

#> # A tibble: 6 × 3
#>    phase  flux std_err
#>    <dbl> <dbl>   <dbl>
#> 1 -0.499 0.938  0.0102
#> 2 -0.498 0.930  0.0102
#> 3 -0.496 0.944  0.0102
#> 4 -0.494 0.963  0.0102
#> 5 -0.494 0.934  0.0102
#> 6 -0.494 0.949  0.0102

cv_tf <- cv_trendfilter(
  x = EB$phase,
  y = EB$flux,
  weights = 1 / EB$std_err^2,
  optimization_params = list(
    max_iter = 1e4,
    obj_tol = 1e-6,
    thinning = TRUE
  )
)

pred_tf <- predict(
  cv_tf,
  validation_error_metric = "MAE",
  lambda_choice = "lambda_1se",
  nx_eval = 1500L
)

#> Error in predict.cv_tf(cv_tf, validation_error_metric = "MAE", lambda_choice = "lambda_1se",     nx_eval = 1500L): validation_error_metric %in% names(obj$validation_errors) is not TRUE