sure_tf class

A list object returned by sure_trendfilter(), with the following elements:

x_eval: Input grid used to evaluate the optimized trend filtering estimate on.

tf_estimate: Optimized trend filtering estimate, evaluated at x_eval.

validation_method: "SURE"

lambdas: Vector of hyperparameter values evaluated in the grid search (always returned in descending order).

edfs: Vector of effective degrees of freedom for all trend filtering estimators fit during validation.

generalization_errors: Vector of SURE generalization error estimates, corresponding to the descending-ordered lambdas vector.

lambda_min: Hyperparameter value that minimizes the SURE generalization error curve.

i_min: Index of lambdas that minimizes the SURE error curve.

edf_min: Effective degrees of freedom of the optimized trend filtering estimator.

n_iter: The number of iterations needed for the ADMM algorithm to converge within the given tolerance, for each hyperparameter value. If many of these are exactly equal to max_iter, then their solutions have not converged with the tolerance specified by obj_tol. In which case, it is often prudent to increase max_iter.

training_errors: Mean-squared error between the observed outputs y and the trend filtering estimate, for every hyperparameter choice.

optimisms: SURE-estimated optimisms, i.e. optimisms = generalization_errors - training_errors.

x: Vector of observed inputs.

y: Vector of observed outputs.

weights: Weights for the observed outputs, defined as the reciprocal variance of the additive noise that contaminates the signal.

fitted_values: Optimized trend filtering estimate, evaluated at the observed inputs x.

residuals: residuals = y - fitted_values

k: Degree of the trend filtering estimator.

admm_params: List of parameter settings for the trend filtering ADMM algorithm, constructed by passing the optimization_params list to glmgen::trendfilter.control.list().

thinning: Logical. If TRUE, then the data are preprocessed so that a smaller, better conditioned data set is used for fitting.

seed: Random number seed (for reproducible results).

x_scale, y_scale, data_scaled: For internal use.

See also