Optimize the trend filtering hyperparameter by minimizing Stein's unbiased risk estimate

For every candidate hyperparameter value, compute an unbiased estimate of the trend filtering model's predictive mean-squared error. See the details section for guidelines on when sure_trendfilter() should be used versus cv_trendfilter().

sure_trendfilter(x, y, weights, k = 2L, nlambdas = 250L, optimization_params)

Arguments

x
Vector of observed values for the input variable.

y
Vector of observed values for the output variable.

weights
(Optional) Weights for the observed outputs, defined as the reciprocal variance of the additive noise that contaminates the output signal. When the noise is expected to have an equal variance, \(\sigma^2\), for all observations, a scalar can be passed to weights, namely weights = \(1/\sigma^2\). Otherwise, weights must be a vector with the same length as x and y.

k
Degree of the polynomials that make up the piecewise-polynomial trend filtering estimate. Defaults to k = 2 (i.e. a piecewise quadratic estimate). Must be one of k = 0,1,2. Higher order polynomials are disallowed since their smoothness is indistinguishable from k = 2 and their use can lead to instability in the convex optimization.

nlambdas
Number of hyperparameter values to test during validation. Defaults to nlambdas = 250. The hyperparameter grid is internally constructed to span the full trend filtering model space lying by a global polynomial solution (i.e. a power law) and an interpolating solution, with nlambdas controlling the granularity of the model space that is

optimization_params
(Optional) A named list of parameter values to be passed to the trend filtering ADMM algorithm of Ramdas and Tibshirani (2016), which is implemented in the glmgen R package. See the glmgen::trendfilter.control.list() documentation for a full list of the algorithm's parameters. The default parameter choices will almost always suffice, but when adjustments do need to be made, one can do so without any technical understanding of the ADMM algorithm. In particular, the four parameter descriptions below should serve as sufficient working knowledge.

obj_tol

A stopping threshold for the ADMM algorithm. If the relative change in the algorithm's cost functional between two consecutive steps is less than obj_tol, the algorithm terminates. The algorithm's termination can also result from it reaching the maximum tolerable iterations set by the max_iter parameter (see below). The obj_tol parameter defaults to obj_tol = 1e-10. The cost_functional vector, returned within the sure_trendfilter() output, gives the relative change in the trend filtering cost functional over the algorithm's final iteration, for every candidate hyperparameter value.

max_iter

Maximum number of ADMM iterations that we will tolerate. Defaults to max_iter = length(y). The actual number of iterations performed by the algorithm, for every candidate hyperparameter value, is returned in the n_iter vector, within the sure_trendfilter() output. If any of the elements of n_iter are equal to max_iter, the tolerance defined by obj_tol has not been attained and max_iter may need to be increased.

thinning

Logical. If thinning = TRUE, then the data are preprocessed so that a smaller data set is used to fit the trend filtering estimate, which will ease the ADMM algorithm's convergence. This can be very useful when a signal is so well-sampled that very little additional information / predictive accuracy is gained by fitting the trend filtering estimate on the full data set, compared to some subset of it. See the cv_trendfilter() examples for a case study of this nature. When nothing is passed to thinning, the algorithm will automatically detect whether thinning should be applied. This preprocessing procedure is controlled by the x_tol parameter below.

x_tol

Controls the automatic detection of when thinning should be applied to the data. If we make bins of size x_tol and find at least two elements of x that fall into the same bin, then the data is thinned.

Value

An object of class 'sure_tf'. This is a list with the following elements:

lambdas

Vector of candidate hyperparameter values (always returned in descending order).

edfs

Number of effective degrees of freedom in the trend filtering estimator, for every hyperparameter value in lambdas.

errors

Vector of mean-squared prediction errors estimated by SURE, for every hyperparameter value in lambdas.

se_errors

Vector of estimated standard errors for the errors.

lambda_min

Hyperparameter value that minimizes the SURE validation error curve.

lambda_1se

The largest hyperparameter value (corresponding to the smoothest trend filtering estimate) that yields a SURE error within one standard error of min(errors). We call this the "1-standard-error rule" hyperparameter, and it serves as an Occam's razor-esque heuristic. That is, given two models with approximately equal performance (here, in terms of MSE), it may be wise to opt for the simpler model, i.e. the model with fewer effective degrees of freedom.

edf_min

Number of effective degrees of freedom in the minimum-SURE trend filtering estimator.

edf_1se

Number of effective degrees of freedom in the 1-stand-error rule trend filtering estimator.

i_min

Index of lambdas that minimizes the SURE error curve.

i_1se

Index of lambdas that gives the 1-standard-error rule hyperparameter.

cost_functional

The relative change in the cost functional over the ADMM algorithm's final iteration, for every candidate hyperparameter in lambdas.

n_iter

Total number of iterations taken by the ADMM algorithm, for every candidate hyperparameter in lambdas. If an element of n_iter is exactly equal to max_iter, then the ADMM algorithm stopped before reaching the tolerance set by obj_tol. In these cases, you may need to increase max_iter to ensure the trend filtering solution has converged to satisfactory precision.

training_errors

In-sample mean-squared error between the observed outputs y and the trend filtering estimate, for every hyperparameter value in lambdas.

optimisms

SURE-estimated optimisms, i.e. optimisms = errors - training_errors.

tf_model

A list of objects that is used internally by other functions that operate on the 'sure_tf' object.

Details

Our recommendations for when to use cv_trendfilter() versus sure_trendfilter() are shown in the table below.

Scenario	Hyperparameter optimization

x is unevenly sampled	cv_trendfilter()

x is evenly sampled and reciprocal variances are not available	cv_trendfilter()

x is evenly sampled and reciprocal variances are available	sure_trendfilter()

For our purposes, an evenly sampled data set with some discarded pixels (either sporadically or in large consecutive chunks) is still considered to be evenly sampled. When the inputs are evenly sampled on a transformed scale, we recommend transforming to that scale and carrying out the full trend filtering analysis on that scale. See the example below for a case when the inputs are evenly sampled on the log10(x) scale.

References

Companion references

Politsch et al. (2020a). Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy. MNRAS, 492(3), p. 4005-4018.

Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical signals with varying degrees of smoothness. MNRAS, 492(3), p. 4019-4032.

Stein's unbiased risk estimate

Tibshirani and Wasserman (2015). Stein’s Unbiased Risk Estimate. 36-702: Statistical Machine Learning course notes (Carnegie Mellon University).

Efron (2014). The Estimation of Prediction Error: Covariance Penalties and Cross-Validation. Journal of the American Statistical Association. 99(467), p. 619-632.

Stein (1981). Estimation of the Mean of a Multivariate Normal Distribution. The Annals of Statistics. 9(6), p. 1135-1151.

Effective degrees of freedom for trend filtering

Tibshirani and Taylor (2012). Degrees of freedom in lasso problems. The Annals of Statistics, 40(2), p. 1198-1232.

See also

Examples

data(quasar_spectrum)
head(spec)

#> # A tibble: 6 × 3
#>   log10_wavelength   flux weights
#>              <dbl>  <dbl>   <dbl>
#> 1             3.55  0.424  0.0417
#> 2             3.55 -2.11   0.125 
#> 3             3.55 -3.78   0.128 
#> 4             3.55  3.60   0.121 
#> 5             3.55 -4.89   0.134 
#> 6             3.55  1.76   0.132 

sure_tf <- sure_trendfilter(spec$log10_wavelength, spec$flux, spec$weights)