The state-of-the-art method for denoising 1D signals

References

Companion references

Politsch et al. (2020a). Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy. MNRAS, 492(3), p. 4005-4018.

Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical signals with varying degrees of smoothness. MNRAS, 492(3), p. 4019-4032.

Trend filtering theory

Tibshirani (2014). Adaptive piecewise polynomial estimation via trend filtering. The Annals of Statistics. 42(1), p. 285-323.

Tibshirani (2020). Divided Differences, Falling Factorials, and Discrete Splines: Another Look at Trend Filtering and Related Problems. arXiv: 2003.03886.

Trend filtering convex optimization algorithm

Ramdas and Tibshirani (2016). Fast and Flexible ADMM Algorithms for Trend Filtering. Journal of Computational and Graphical Statistics, 25(3), p. 839-858.

Arnold, Sadhanala, and Tibshirani (2014). Fast algorithms for generalized lasso problems. R package glmgen. Version 0.0.3.

Effective degrees of freedom for trend filtering

Tibshirani and Taylor (2012). Degrees of freedom in lasso problems. The Annals of Statistics, 40(2), p. 1198-1232.

Stein's unbiased risk estimate

Tibshirani and Wasserman (2015). Stein’s Unbiased Risk Estimate. 36-702: Statistical Machine Learning course notes (Carnegie Mellon University).

Efron (2014). The Estimation of Prediction Error: Covariance Penalties and Cross-Validation. Journal of the American Statistical Association. 99(467), p. 619-632.

Stein (1981). Estimation of the Mean of a Multivariate Normal Distribution. The Annals of Statistics. 9(6), p. 1135-1151.

The Bootstrap and variations

Efron and Tibshirani (1986). Bootstrap Methods for Standard Errors, Confidence Intervals, and Other Measures of Statistical Accuracy. Statistical Science, 1(1), p. 54-75.

Mammen (1993). Bootstrap and Wild Bootstrap for High Dimensional Linear Models. The Annals of Statistics, 21(1), p. 255-285.

Efron (1979). Bootstrap Methods: Another Look at the Jackknife. The Annals of Statistics, 7(1), p. 1-26.

Cross validation

Hastie, Tibshirani, and Friedman (2009). The Elements of Statistical Learning: Data Mining, Inference, and Prediction. 2nd edition. Springer Series in Statistics. (See Sections 7.10 and 7.12)

James, Witten, Hastie, and Tibshirani (2013). An Introduction to Statistical Learning: with Applications in R. Springer. (See Section 5.1; Less technical than ESL)

Tibshirani (2013). Model selection and validation 2: Model assessment, more cross-validation. 36-462: Data Mining course notes (Carnegie Mellon University).

Author

Collin A. Politsch, Ph.D.

Email: collinpolitsch@gmail.com
Website: collinpolitsch.com
GitHub: github.com/capolitsch