Construct pointwise variability bands via a tailored bootstrap algorithm

Generate a bootstrap ensemble of trend filtering estimates in order to quantify the uncertainty in the optimized estimate. One of three possible bootstrap algorithms should be chosen according to the criteria in the Details section below. Pointwise variability bands are then obtained by passing the 'bootstrap_trendfilter' object to vbands(), along with the desired level (e.g. level = 0.95) .

bootstrap_trendfilter(
  obj,
  algorithm = NULL,
  B = 100L,
  x_eval = NULL,
  lambda = NULL,
  mc_cores = parallel::detectCores() - 4,
  ...
)

Arguments

obj
An object of class 'cv_trendfilter' or 'sure_trendfilter'.

algorithm
A string specifying which variation of the bootstrap to use. One of c("nonparametric", "parametric", "wild"). See Details section below for guidelines on when each choice should be used.

B
The number of bootstrap samples used to estimate the pointwise variability bands. Defaults to B = 100L.

mc_cores
Number of cores to utilize for parallel computing. Defaults to the number of cores detected, minus 4.

...
Additional named arguments. Currently unused.

Value

An object of class 'bootstrap_trendfilter' and subclass 'trendfilter'. This is a list with the elements below, as well as all elements from obj.

x_eval

Input grid that each bootstrap trend filtering estimate was evaluated on.

ensemble

The full trend filtering bootstrap ensemble as a matrix with length(x_eval) rows and B columns.

edf_boots

Vector of the estimated number of effective degrees of freedom of each trend filtering bootstrap estimate.

n_iter_boots

Vector of the number of iterations taken by the ADMM algorithm before reaching a stopping criterion, for each bootstrap estimate.

lambda_boots

Vector of the hyperparameter values used for each bootstrap fit. In general, these are not all equal because our bootstrap implementation instead seeks to hold the number of effective degrees of freedom constant across all bootstrap estimates.

algorithm

A string specifying which variation of the bootstrap was used to generate the ensemble.

Details

Our recommendations for when to use each of the possible settings for the algorithm argument are shown in the table below. See Politsch et al. (2020a) for more details.

Scenario	Uncertainty quantification

x is unevenly sampled	algorithm = "nonparametric"

x is evenly sampled and measurement variances for y are available	algorithm = "parametric"

x is evenly sampled and measurement variances for y are not available	algorithm = "wild"

For our purposes, an evenly sampled data set with some discarded pixels (either sporadically or in large consecutive chunks) is still considered to be evenly sampled. When the inputs are evenly sampled on a transformed scale, we recommend transforming to that scale and carrying out the full trend filtering analysis on that scale. See Example 2 below for a case when the inputs are evenly sampled on the log10(x) scale.

References

Politsch et al. (2020a). Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy. MNRAS, 492(3), p. 4005-4018. [Publisher] [arXiv].

Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical signals with varying degrees of smoothness. MNRAS, 492(3), p. 4019-4032. [Publisher] [arXiv].

See also

Examples

# Example 1: Phase-folded light curve of an eclipsing binary star system
#
# The apparent brightness over time of a star system that has two suns
# that regularly eclipse one another from our vantage point on Earth. Here,
# the time series is stacked according to the orbital period of the binary
# system, with the primary eclipse occuring at `phase = 0` and the input
# domain ranging from -0.5 to 0.5.

data("eclipsing_binary")
head(eclipsing_binary)

#> # A tibble: 6 × 3
#>    phase  flux std_err
#>    <dbl> <dbl>   <dbl>
#> 1 -0.499 0.938  0.0102
#> 2 -0.498 0.930  0.0102
#> 3 -0.496 0.944  0.0102
#> 4 -0.494 0.963  0.0102
#> 5 -0.494 0.934  0.0102
#> 6 -0.494 0.949  0.0102

x <- eclipsing_binary$phase
y <- eclipsing_binary$flux
weights <- 1 / eclipsing_binary$std_err^2

cv_tf <- cv_trendfilter(x, y, weights, max_iter = 1e4, obj_tol = 1e-6)
boot_tf <- bootstrap_trendfilter(cv_tf, "nonparametric")

#> Warning: all scheduled cores encountered errors in user code

#> Error in par_out[[X]][["tf_estimate_boot"]]: subscript out of bounds

# Example 2: The "Lyman-alpha forest" in the spectrum of a distant quasar

data("quasar_spectrum")
head(quasar_spectrum)

#> # A tibble: 6 × 3
#>   log10_wavelength   flux weights
#>              <dbl>  <dbl>   <dbl>
#> 1             3.55  0.424  0.0417
#> 2             3.55 -2.11   0.125 
#> 3             3.55 -3.78   0.128 
#> 4             3.55  3.60   0.121 
#> 5             3.55 -4.89   0.134 
#> 6             3.55  1.76   0.132 

x <- quasar_spectrum$log10_wavelength
y <- quasar_spectrum$flux
weights <- quasar_spectrum$weights

sure_tf <- sure_trendfilter(x, y, weights)
boot_tf <- bootstrap_trendfilter(sure_tf, "parametric")