Construct pointwise variability bands via a bootstrap

Construct pointwise variability bands via a bootstrap

bootstrap_trendfilter(
  obj,
  bootstrap_algorithm,
  level = 0.95,
  B = 100L,
  return_ensemble = FALSE,
  prune = TRUE,
  mc_cores = parallel::detectCores()
)

Arguments

obj
An object of class sure_tf or cv_tf.

bootstrap_algorithm
A string specifying which variation of the bootstrap to use. One of c("nonparametric","parametric","wild"). See details below for recommendations on when each option is appropriate.

level
The level of the pointwise variability bands. Defaults to level = 0.95.

B
The number of bootstrap samples used to estimate the pointwise variability bands. Defaults to B = 100.

return_ensemble
Logical. If TRUE, the full trend filtering bootstrap ensemble is returned as an \(n \times B\) matrix, less any columns from post-hoc pruning (see prune below). Defaults to return_ensemble = FALSE to save memory.

prune
Logical. If TRUE, then the trend filtering bootstrap ensemble is examined for rare instances in which the optimization has stopped at zero knots (likely erroneously), and removes them from the ensemble that is used to compute the variability bands. Defaults to prune = TRUE. Do not change this unless you know what you are doing!

mc_cores
Multi-core computing using the parallel package: The number of cores to utilize. Defaults to the number of cores detected.

Value

An object of class bootstrap_tf. This is a comprehensive list containing all of the analysis important information, data, and results:

tf_standard_errors

The standard errors of the optimized trend filtering point estimator.

bootstrap_lower_band

Vector of lower bounds for the pointwise variability bands, evaluated on x_eval.

bootstrap_upper_band

Vector of upper bounds for the pointwise variability bands, evaluated on x_eval.

bootstrap_algorithm

A string specifying which variation of the bootstrap was used to obtain the variability bands.

level

The level of the pointwise variability bands.

B

The number of bootstrap samples used to estimate the pointwise variability bands.

tf_bootstrap_ensemble

If return_ensemble = TRUE, the full trend filtering bootstrap ensemble as an \(n \times B\) matrix, less any columns from post-hoc pruning (if prune = TRUE). Else, this will return NULL.

edf_boots

An integer vector of the estimated number of effective degrees of freedom of each trend filtering bootstrap estimate. These should all be relatively close to edf_min (below).

prune

Logical. If TRUE, then the trend filtering bootstrap ensemble is examined for rare instances in which the optimization has stopped at zero knots (likely erroneously), and removes them from the ensemble.

n_pruned

The number of poorly-converged bootstrap trend filtering estimates pruned from the ensemble.

n_iter_boots

Vector of the number of iterations needed for the ADMM algorithm to converge within the given tolerance, for each bootstrap trend filtering estimate.

...

Named elements inherited from obj --- an object either of class sure_tf or cv_tf. See the relevant function documentation for details.

Details

Our recommendations for when to use each of the possible settings for the bootstrap_algorithm argument are shown in the table below. See Politsch et al. (2020a) for more details.

Scenario	Uncertainty quantification

x is irregularly sampled	bootstrap_algorithm = "nonparametric"

x is regularly sampled and reciprocal variances are not available	bootstrap_algorithm = "wild"

x is regularly sampled and reciprocal variances are available	bootstrap_algorithm = "parametric"

References

Companion references

Politsch et al. (2020a). Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy. MNRAS, 492(3), p. 4005-4018.

Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical signals with varying degrees of smoothness. MNRAS, 492(3), p. 4019-4032.

The Bootstrap and variations

Efron and Tibshirani (1986). Bootstrap Methods for Standard Errors, Confidence Intervals, and Other Measures of Statistical Accuracy. Statistical Science, 1(1), p. 54-75.

Mammen (1993). Bootstrap and Wild Bootstrap for High Dimensional Linear Models. The Annals of Statistics, 21(1), p. 255-285.

Efron (1979). Bootstrap Methods: Another Look at the Jackknife. The Annals of Statistics, 7(1), p. 1-26.

See also

Examples

data(quasar_spectrum)
head(spec)

#> # A tibble: 6 × 3
#>   log10_wavelength   flux weights
#>              <dbl>  <dbl>   <dbl>
#> 1             3.55  0.424  0.0417
#> 2             3.55 -2.11   0.125 
#> 3             3.55 -3.78   0.128 
#> 4             3.55  3.60   0.121 
#> 5             3.55 -4.89   0.134 
#> 6             3.55  1.76   0.132 

if (FALSE) {
sure_tf <- sure_trendfilter(spec$log10_wavelength, spec$flux, spec$weights)
opt_tf <- bootstrap_trendfilter(sure_tf, bootstrap_algorithm = "parametric")
}