Optimize the trend filtering hyperparameter by V-fold cross validation

For every candidate hyperparameter value, estimate the trend filtering model's out-of-sample error by V-fold cross validation. CV error curves are returned for four of the most common regression loss metrics, as well as observation-weighted versions. Custom loss functions may also be passed to the loss_funcs argument. See the details section for definitions of the internal loss functions, and for guidelines on when cv_trendfilter() should be used versus sure_trendfilter().

cv_trendfilter(
  x,
  y,
  weights,
  k = 2L,
  nlambdas = 250L,
  V = 10L,
  mc_cores = V,
  loss_funcs,
  fold_ids,
  optimization_params
)

Arguments

x
Vector of observed values for the input variable.

y
Vector of observed values for the output variable.

weights
(Optional) Weights for the observed outputs, defined as the reciprocal variance of the additive noise that contaminates the output signal. When the noise is expected to have an equal variance \(\sigma^2\) for all observations, a scalar may be passed to weights, namely weights = \(1/\sigma^2\). Otherwise, weights must be a vector with the same length as x and y.

k
Degree of the polynomials that make up the piecewise-polynomial trend filtering estimate. Defaults to k = 2 (i.e. a piecewise quadratic estimate). Must be one of k = 0,1,2. Higher order polynomials are disallowed since their smoothness is indistinguishable from k = 2 and their use can lead to instability in the convex optimization.

nlambdas
Number of hyperparameter values to test during validation. Defaults to nlambdas = 250. The hyperparameter grid is internally constructed to span the full trend filtering model space (which is bookended by a global polynomial solution and an interpolating solution), with nlambdas controlling the granularity of the hyperparameter grid.

V
Number of folds that the data are partitioned into for V-fold cross validation. Defaults to V = 10.

mc_cores
Multi-core computing using the parallel R package: The number of cores to utilize. Defaults to the number of cross validation folds, mc_cores = V. If the value passed to mc_cores exceeds the number of cores available on the machine, mc_cores is internally updated to mc_cores = parallel::detectCores().

loss_funcs
(Optional) A named list of one or more functions, with each defining a loss function to be evaluated on held-out folds during cross validation. By default, cv_trendfilter() will automatically compute and return CV error curves for 9 common regression loss functions (see the Details section below). Therefore, the loss_funcs argument need only be used in order to define loss functions that are not among these 9 choices.

When defining custom loss functions, each function within the named list passed to loss_funcs should take three vector arguments --- y, tf_estimate, and weights --- and return a single scalar value for the validation error. For example, if I wanted a CV error curve based on a weighted median of the absolute errors to be computed, I would pass the list below to loss_funcs:

MedAE <- function(tf_estimate, y, weights) {
  matrixStats::weightedMedian(abs(tf_estimate - y), sqrt(weights))
}

list(MedAE = MedAE)

fold_ids
(Optional) An integer vector defining a custom partition of the data for cross validation. fold_ids must have the same length as x and y, and only contain values in the set 1:V.

optimization_params
(Optional) A named list of parameter values to be passed to the trend filtering ADMM algorithm of Ramdas and Tibshirani (2016), which is implemented in the glmgen R package. See the glmgen::trendfilter.control.list() documentation for a full list of the algorithm's parameters. The default parameter choices will almost always suffice, but when adjustments do need to be made, one can do so without any technical understanding of the ADMM algorithm. In particular, the four parameter descriptions below should serve as sufficient working knowledge.

obj_tol

A stopping threshold for the ADMM algorithm. If the relative change in the algorithm's cost functional between two consecutive steps is less than obj_tol, the algorithm terminates. The algorithm's termination can also result from it reaching the maximum tolerable iterations set by the max_iter parameter below. The obj_tol parameter defaults to obj_tol = 1e-10. The cost_functional vector returned within the sure_trendfilter() output gives the relative change in the trend filtering cost functional over the algorithm's final iteration, for every candidate hyperparameter value.

max_iter

Maximum number of ADMM iterations that we will tolerate. Defaults to max_iter = length(y). The actual number of iterations performed by the algorithm, for every candidate hyperparameter value, is returned in the n_iter vector within the sure_trendfilter() output. If any of the elements of n_iter are equal to max_iter, the tolerance defined by obj_tol has not been attained and max_iter may need to be increased.

thinning

Logical. If thinning = TRUE, then the data are preprocessed so that a smaller data set is used to fit the trend filtering estimate, which will ease the ADMM algorithm's convergence. This can be very useful when, for example, a signal is so well-sampled that very little additional information / predictive accuracy is gained by fitting the trend filtering estimate on the full data set, compared to only fitting on some subset of it. See the cv_trendfilter() examples for a case study of this nature. When nothing is passed to thinning, the algorithm will automatically detect whether thinning should be applied. This preprocessing procedure is controlled by the x_tol parameter below.

x_tol

Controls the automatic detection of when thinning should be applied to the data. If we make bins of size x_tol and find at least two elements of x that fall within the same bin, then the data is thinned.

Value

An object of class 'cv_tf'. This is a list with the following elements:

lambdas

Vector of candidate hyperparameter values (always returned in descending order).

edfs

Number of effective degrees of freedom in the trend filtering estimator, for every candidate hyperparameter value in lambdas.

errors

A named list of vectors, with each representing the CV error curve for every loss function in loss_funcs (see below).

se_errors

Standard errors for the errors, within a named list of the same structure.

lambda_min

A named vector with length equal to length(errors), containing the hyperparameter value that minimizes the CV error curve, for every loss function in loss_funcs.

lambda_1se

A named vector with length equal to length(errors), containing the "1-standard-error rule" hyperparameter, for every loss function in loss_funcs. The "1-standard-error rule" hyperparameter is the largest hyperparameter value in lambdas (corresponding to the smoothest trend filtering estimate) that has a CV error within one standard error of min(errors). It serves as an Occam's razor-like heuristic. More precisely, given two models with approximately equal performance (in terms of some loss function), it may be wise to opt for the simpler model, i.e. the model with the larger hyperparameter / fewer effective degrees of freedom.

edf_min

A named vector with length equal to length(errors), containing the number of effective degrees of freedom in the trend filtering estimator that minimizes the CV error curve, for every loss function in loss_funcs.

edf_1se

A named vector with length equal to length(errors), containing the number of effective degrees of freedom in the "1-standard-error rule" trend filtering estimator, for every loss function in loss_funcs.

i_min

A named vector with length equal to length(errors), containing the index of lambdas that minimizes the CV error curve, for every loss function in loss_funcs.

i_1se

A named vector with length equal to length(errors), containing the index of lambdas that gives the "1-standard-error rule" hyperparameter value, for every loss function in loss_funcs..

loss_funcs

A named list of functions that defines all loss functions evaluated during cross validation.

cost_functional

The relative change in the cost functional over the ADMM algorithm's final iteration, for every candidate hyperparameter in lambdas.

n_iter

Total number of iterations taken by the ADMM algorithm, for every candidate hyperparameter in lambdas. If an element of n_iter is exactly equal to tf_model$admm_params$max_iter, then the ADMM algorithm stopped before reaching the objective tolerance tf_model$admm_params$obj_tol. In these situations, you may need to increase the maximum number of tolerable iterations via the optimization_params$max_iter argument of cv_trendfilter() in order to ensure that the trend filtering solution has converged to satisfactory precision.

V

The number of folds the data were split into for cross validation.

tf_model

A list of objects that is used internally by other functions that operate on the cv_trendfilter() output.

Details

Our recommendations for when to use cv_trendfilter() versus sure_trendfilter() are summarized in the table below. See Section 3.5 of Politsch et al. (2020a) for more details.

Scenario	Hyperparameter optimization

x is unevenly sampled	cv_trendfilter()

x is evenly sampled and reciprocal variances are not available	cv_trendfilter()

x is evenly sampled and reciprocal variances are available	sure_trendfilter()

For our purposes, an evenly sampled data set with some discarded pixels (either sporadically or in large consecutive chunks) is still considered to be evenly sampled. When x is evenly sampled on a transformed scale, we recommend transforming to that scale and carrying out the full trend filtering analysis on that scale. See the sure_trendfilter() examples for a case when the inputs are evenly sampled on the log10(x) scale.

The following loss functions are automatically computed during cross validation and their CV error curves are returned within the errors list of the 'cv_tf' output object.

Mean absolute deviations error: \[\text{MAE}(\lambda) = \frac{1}{n} \sum_{i=1}^{n}|Y_i - \hat{f}(x_i; \lambda)|\]

Weighted mean absolute deviations error: \[\text{WMAE}(\lambda) = \sum_{i=1}^{n} |Y_i - \hat{f}(x_i; \lambda)|\frac{\sqrt{w_i}}{\sum_j\sqrt{w_j}}\]

Mean-squared error: \[\text{MSE}(\lambda) = \frac{1}{n} \sum_{i=1}^{n} |Y_i - \hat{f}(x_i; \lambda)|^2\]

Weighted mean-squared error: \[\text{WMSE}(\lambda) = \sum_{i=1}^{n}|Y_i - \hat{f}(x_i; \lambda)|^2\frac{w_i}{\sum_jw_j}\]

log-cosh error: \[\text{logcosh}(\lambda) = \frac{1}{n} \sum_{i=1}^{n} \log\left(\cosh\left(Y_i - \hat{f}(x_i; \lambda)\right)\right)\]

Weighted log-cosh error: \[\text{wlogcosh}(\lambda) = \sum_{i=1}^{n} \log\left(\cosh\left((Y_i - \hat{f}(x_i; \lambda))\sqrt{w_i}\right)\right)\]

Huber loss: \[\text{Huber}(\lambda) = \frac{1}{n}\sum_{i=1}^{n}L_{\delta}(Y_i; \lambda)\] \[\text{where}\;\;\;\;L_{\delta}(Y_i; \lambda) = \cases{ |Y_i - \hat{f}(x_i; \lambda)|^2, & $|Y_i - \hat{f}(x_i; \lambda)| \leq \delta$ \cr 2\delta|Y_i - \hat{f}(x_i; \lambda)| - \delta^2, & $|Y_i - \hat{f}(x_i; \lambda)| > \delta$}\]

Weighted Huber loss: \[\text{wHuber}(\lambda) = \sum_{i=1}^{n}L_{\delta}(Y_i; \lambda)\] \[\text{where}\;\;\;\;L_{\delta}(Y_i; \lambda) = \cases{ |Y_i - \hat{f}(x_i; \lambda)|^2w_i, & $|Y_i - \hat{f}(x_i; \lambda)|\sqrt{w_i} \leq \delta$ \cr 2\delta|Y_i - \hat{f}(x_i; \lambda)|\sqrt{w_i} - \delta^2, & $|Y_i - \hat{f}(x_i; \lambda)|\sqrt{w_i} > \delta$}\]

Mean-squared logarithmic error: \[\text{MSLE}(\lambda) = \frac{1}{n}\sum_{i=1}^{n} \left|\log(Y_i + 1) - \log(\hat{f}(x_i; \lambda) + 1)\right|\]

where \(w_i:=\) weights[i].

References

Companion references

Politsch et al. (2020a). Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy. MNRAS, 492(3), p. 4005-4018. [arXiv].

Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical signals with varying degrees of smoothness. MNRAS, 492(3), p. 4019-4032. [arXiv].

Cross validation

Hastie, Tibshirani, and Friedman (2009). The Elements of Statistical Learning: Data Mining, Inference, and Prediction. 2nd edition. Springer Series in Statistics. (See Sections 7.10 and 7.12).

See also

Examples

data(eclipsing_binary)
head(EB)

#> # A tibble: 6 × 3
#>    phase  flux std_err
#>    <dbl> <dbl>   <dbl>
#> 1 -0.499 0.938  0.0102
#> 2 -0.498 0.930  0.0102
#> 3 -0.496 0.944  0.0102
#> 4 -0.494 0.963  0.0102
#> 5 -0.494 0.934  0.0102
#> 6 -0.494 0.949  0.0102

cv_tf <- cv_trendfilter(
  x = EB$phase,
  y = EB$flux,
  weights = 1 / EB$std_err^2,
  optimization_params = list(
    max_iter = 1e4,
    obj_tol = 1e-6,
    thinning = TRUE
  )
)