Optimize the trend filtering hyperparameter by V-fold cross validation

cv_trendfilter() optimizes the trend filtering hyperparameter via V-fold cross validation on a grid of candidate hyperparameter settings and selects the value that minimizes a user-specified loss metric. See details for when to use sure_trendfilter() vs. cv_trendfilter().

cv_trendfilter(
  x,
  y,
  weights,
  k = 2L,
  nlambdas = 250L,
  lambdas,
  V = 10L,
  lambda_choice = c("lambda_min", "lambda_1se"),
  validation_functional = "WMAE",
  nx_eval = 1500L,
  x_eval,
  mc_cores = parallel::detectCores(),
  optimization_params
)

Arguments

x
Vector of observed values for the input variable.

y
Vector of observed values for the output variable.

weights
(Optional) Weights for the observed outputs, defined as the reciprocal variance of the additive noise that contaminates the signal. weights can be passed as a scalar when the noise is expected to have equal variance for all observations. Otherwise, weights must have the same length as x and y.

k
Degree of the piecewise polynomials that make up the trend filtering estimate. Defaults to k = 2 (i.e. a piecewise quadratic estimate). Must be one of k = 0,1,2. Higher order polynomials are disallowed since their smoothness is indistinguishable from k = 2 and their use can lead to instability in the convex optimization.

nlambdas
The number of hyperparameter settings to test during validation. When nothing is passed to lambdas (highly recommended for general use), the grid is automatically constructed by cv_trendfilter(), with nlambdas controlling the granularity of the grid.

lambdas
(Optional) Overrides nlambdas if passed. The vector of trend filtering hyperparameter values for the grid search. Use of this argument is discouraged unless you know what you are doing.

V
Number of folds the data are partitioned into for the V-fold cross validation. Defaults to V = 10.

lambda_choice
One of c("lambda_min","lambda_1se"). The choice of hyperparameter that is used for optimized trend filtering estimate. Defaults to lambda_choice = "lambda_min".

"lambda_min": The hyperparameter value that minimizes the cross validation error curve.

"lambda_1se": The largest hyperparameter value with a cross validation error within 1 standard error of the minimum cross validation error. This choice therefore favors simpler (i.e. smoother) trend filtering estimates. The motivation here is essentially Occam's razor: the two models yield results that are quantitatively very close, so we favor the simpler model. See Section 7.10 of Hastie, Tibshirani, and Friedman (2009) for more details on the ``one-standard-error rule''.

validation_functional
Loss functional to optimize during cross validation. Some common choices can be used by passing an appropriate string --- one of c("MSE","MAE","WMSE","WMAE"), i.e. mean-absolute deviations error, mean-squared error, and their weighted counterparts. Defaults to validation_functional = "WMAE".

Alternatively, custom validation loss functionals can be used by instead passing a function to validation_functional. The function should take three vector arguments --- y, tf_estimate, and weights --- and return a single scalar value for the validation loss. For example, validation_functional = "WMAE" is equivalent to passing the following function:

function(tf_estimate, y, weights){
  sum(abs(tf_estimate - y) * sqrt(weights) / sum(sqrt(weights)))
}

nx_eval
Integer. If nothing is passed to x_eval, then it is defined as x_eval = seq(min(x), max(x), length = nx_eval).

x_eval
(Optional) A grid of inputs to evaluate the optimized trend filtering estimate on. May be ignored, in which case the grid is determined by nx_eval.

mc_cores
Multi-core computing using the parallel package: The number of cores to utilize. Defaults to the number of cores detected.

optimization_params
(Optional) A named list of parameter choices to be passed to the trend filtering ADMM algorithm (Ramdas and Tibshirani 2016). See the glmgen::trendfilter.control.list() documentation for full details. No technical understanding of the ADMM algorithm is needed and the default parameter choices will almost always suffice. However, the following parameters may require some adjustments to ensure that your trend filtering estimate has sufficiently converged:

max_iter: Maximum iterations allowed for the trend filtering convex optimization. Defaults to max_iter = 600L. See the n_iter element of the function output for the actual number of iterations taken for every hyperparameter choice in lambdas. If any of the elements of n_iter are equal to max_iter, the objective function's tolerance has not been achieved and max_iter may need to be increased.

obj_tol: The tolerance used in the convex optimization stopping criterion; when the relative change in the objective function is less than this value, the algorithm terminates. Thus, decreasing this setting will increase the precision of the solution returned by the optimization. Defaults to obj_tol = 1e-10. If the returned trend filtering estimate does not appear to have fully converged to a reasonable estimate of the signal, this issue can be resolve by some combination of decreasing obj_tol and increasing max_iter.

thinning: Logical. If TRUE, then the data are preprocessed so that a smaller, better conditioned data set is used for fitting. When left NULL (the default setting), the optimization will automatically detect whether thinning should be applied (i.e. cases in which the numerical fitting algorithm will struggle to converge). This preprocessing procedure is controlled by the x_tol argument below.

x_tol: Controls the automatic detection of when thinning should be applied to the data. If we make bins of size x_tol and find at least two elements of x that fall into the same bin, then we thin the data.

Value

An object of class cv_tf. This is a list with the following elements:

x_eval

Input grid used to evaluate the optimized trend filtering estimate on.

tf_estimate

Optimized trend filtering estimate, evaluated at x_eval.

validation_method

paste0(V,"-fold CV")

validation_functional

Type of error that validation was performed on. Either one of c("MSE","MAE","WMSE","WMAE") or a custom function passed by the user.

V

The number of folds the data are split into for the V-fold cross validation.

lambdas

Vector of hyperparameter values evaluated in the grid search (always returned in descending order).

edfs

Vector of effective degrees of freedom for all trend filtering estimators fit during validation.

generalization_errors

Vector of cross validation estimates of the trend filtering generalization error, for each hyperparameter value (ordered corresponding to the descending-ordered lambdas vector).

se_errors

The standard errors of the cross validation errors. These are particularly useful for implementing the ``1-standard-error rule''.

lambda_min

Hyperparameter value that minimizes the cross validation generalization error curve.

lambda_1se

Largest hyperparameter value that is within one standard error of the minimum hyperparameter's cross validation error.

lambda_choice

One of c("lambda_min", "lambda_1se"). The choice of hyperparameter that is used for the returned trend filtering estimate evaluation `tf_estimate`.

i_min

Index of `lambdas` that minimizes the cross validation error.

i_1se

Index of `lambdas` that gives the largest hyperparameter value that has a cross validation error within 1 standard error of the minimum of the cross validation error curves.

edf_min

Effective degrees of freedom of the optimized trend filtering estimator.

edf_1se

Effective degrees of freedom of the 1-stand-error rule trend filtering estimator.

n_iter

The number of iterations needed for the ADMM algorithm to converge within the given tolerance, for each hyperparameter value. If many of these are exactly equal to `max_iter`, then their solutions have not converged with the tolerance specified by `obj_tol`. In which case, it is often prudent to increase `max_iter`.

x

Vector of observed inputs.

y

Vector of observed outputs.

weights

Weights for the observed outputs, defined as the reciprocal variance of the additive noise that contaminates the signal.

fitted_values

Optimized trend filtering estimate, evaluated at the observed inputs `x`.

residuals

`residuals = y - fitted_values`

k

Degree of the trend filtering estimator.

admm_params

List of parameter settings for the trend filtering ADMM algorithm, constructed by passing the `optimization_params` list to glmgen::trendfilter.control.list().

thinning

Logical. If `TRUE`, then the data are preprocessed so that a smaller, better conditioned data set is used for fitting.

x_scale, y_scale, data_scaled

For internal use.

Details

Our recommendations for when to use cv_trendfilter() vs. sure_trendfilter() are shown in the table below.

A regularly-sampled data set with some discarded pixels (either sporadically or in large consecutive chunks) is still considered regularly sampled. When the inputs are regularly sampled on a transformed scale, we recommend transforming to that scale and carrying out the full trend filtering analysis on that scale. See the example below for a case when the inputs are evenly sampled on the log10(x) scale.

Scenario	Hyperparameter optimization

x is irregularly sampled	cv_trendfilter()

x is regularly sampled and reciprocal variances are not available	cv_trendfilter()

x is regularly sampled and reciprocal variances are available	sure_trendfilter()

The formal definitions of the common validation loss functionals available via the options validation_functional = c("MSE","MAE","WMSE","WMAE") are stated below.

\[MSE(\lambda) = \frac{1}{n} \sum_{i=1}^{n} |Y_i - \hat{f}(x_i; \lambda)|^2\] \[MAE(\lambda) = \frac{1}{n} \sum_{i=1}^{n} |Y_i - \hat{f}(x_i; \lambda)|\] \[WMSE(\lambda) = \sum_{i=1}^{n} |Y_i - \hat{f}(x_i; \lambda)|^2\frac{w_i}{\sum_jw_j}\] \[WMAE(\lambda) = \sum_{i=1}^{n} |Y_i - \hat{f}(x_i; \lambda)|\frac{\sqrt{w_i}}{\sum_j\sqrt{w_j}}\] where \(w_i\) is the \(i\)th element of the weights vector.

If constant weights are passed, or if nothing is passed, then the weighted and unweighted counterparts are equivalent.

Briefly stated, weighting helps combat heteroskedasticity (varying levels of uncertainty in the output measurements) and absolute error is less sensitive to outliers than squared error.

References

Companion references

Politsch et al. (2020a). Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy. MNRAS, 492(3), p. 4005-4018.

Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical signals with varying degrees of smoothness. MNRAS, 492(3), p. 4019-4032.

Cross validation

Hastie, Tibshirani, and Friedman (2009). The Elements of Statistical Learning: Data Mining, Inference, and Prediction. 2nd edition. Springer Series in Statistics. (See Sections 7.10 and 7.12)

See also

Examples

data(eclipsing_binary)
head(EB)

#> # A tibble: 6 × 3
#>    phase  flux std_err
#>    <dbl> <dbl>   <dbl>
#> 1 -0.499 0.938  0.0102
#> 2 -0.498 0.930  0.0102
#> 3 -0.496 0.944  0.0102
#> 4 -0.494 0.963  0.0102
#> 5 -0.494 0.934  0.0102
#> 6 -0.494 0.949  0.0102

if (FALSE) {
cv_tf <- cv_trendfilter(
  x = EB$phase,
  y = EB$flux,
  weights = 1 / EB$std_err^2,
  validation_functional = "MAE",
  lambdas = exp(seq(20, 7, length = 250)),
  optimization_params = list(
    max_iter = 5e3,
    obj_tol = 1e-6,
    thinning = T
  )
)
}