Optimize the trend filtering hyperparameter by minimizing Stein's unbiased risk estimate

sure_trendfilter() optimizes the trend filtering hyperparameter via a grid search over a vector of candidate hyperparameter settings and selects the value that minimizes an unbiased estimate of the model's generalization error. See details for when to use sure_trendfilter() vs. cv_trendfilter().

sure_trendfilter(
  x,
  y,
  weights,
  k = 2L,
  nlambdas = 250L,
  lambdas,
  nx_eval = 1500L,
  x_eval,
  optimization_params
)

Arguments

x
Vector of observed values for the input variable.

y
Vector of observed values for the output variable.

weights
Weights for the observed outputs, defined as the reciprocal variance of the additive noise that contaminates the signal. weights can be passed as a scalar when the noise is expected to have equal variance for all observations. Otherwise, weights must have the same length as x and y.

k
Degree of the piecewise polynomials that make up the trend filtering estimate. Defaults to k = 2 (i.e. a piecewise quadratic estimate). Must be one of k = 0,1,2. Higher order polynomials are disallowed since their smoothness is indistinguishable from k = 2 and their use can lead to instability in the convex optimization.

nlambdas
The number of hyperparameter settings to test during validation. When nothing is passed to lambdas (highly recommended for general use), the grid is automatically constructed by sure_trendfilter(), with nlambdas controlling the granularity of the grid.

lambdas
(Optional) Overrides nlambdas if passed. The vector of trend filtering hyperparameter values for the grid search. Use of this argument is discouraged unless you know what you are doing.

nx_eval
Integer. If nothing is passed to x_eval, then it is defined as x_eval = seq(min(x), max(x), length = nx_eval).

x_eval
(Optional) A grid of inputs to evaluate the optimized trend filtering estimate on. May be ignored, in which case the grid is determined by nx_eval.

optimization_params
(Optional) A named list of parameter choices to be passed to the trend filtering ADMM algorithm (Ramdas and Tibshirani 2016). See the glmgen::trendfilter.control.list() documentation for full details. No technical understanding of the ADMM algorithm is needed and the default parameter choices will almost always suffice. However, the following parameters may require some adjustments to ensure that your trend filtering estimate has sufficiently converged:

max_iter: Maximum iterations allowed for the trend filtering convex optimization. Defaults to max_iter = 600L. See the n_iter element of the function output for the actual number of iterations taken for every hyperparameter choice in lambdas. If any of the elements of n_iter are equal to max_iter, the objective function's tolerance has not been achieved and max_iter may need to be increased.

obj_tol: The tolerance used in the convex optimization stopping criterion; when the relative change in the objective function is less than this value, the algorithm terminates. Thus, decreasing this setting will increase the precision of the solution returned by the optimization. Defaults to obj_tol = 1e-10. If the returned trend filtering estimate does not appear to have fully converged to a reasonable estimate of the signal, this issue can be resolve by some combination of decreasing obj_tol and increasing max_iter.

thinning: Logical. If TRUE, then the data are preprocessed so that a smaller, better conditioned data set is used for fitting. When left NULL (the default setting), the optimization will automatically detect whether thinning should be applied (i.e. cases in which the numerical fitting algorithm will struggle to converge). This preprocessing procedure is controlled by the x_tol argument below.

x_tol: Controls the automatic detection of when thinning should be applied to the data. If we make bins of size x_tol and find at least two elements of x that fall into the same bin, then we thin the data.

Value

An object of class sure_tf. This is a list with the following elements:

x_eval

Input grid used to evaluate the optimized trend filtering estimate on.

tf_estimate

Optimized trend filtering estimate, evaluated at x_eval.

validation_method

"SURE"

lambdas

Vector of hyperparameter values evaluated in the grid search (always returned in descending order).

edfs

Vector of effective degrees of freedom for all trend filtering estimators fit during validation.

generalization_errors

Vector of SURE generalization error estimates, corresponding to the descending-ordered lambdas vector.

lambda_min

Hyperparameter value that minimizes the SURE generalization error curve.

i_min

Index of lambdas that minimizes the SURE error curve.

edf_min

Effective degrees of freedom of the optimized trend filtering estimator.

n_iter

The number of iterations needed for the ADMM algorithm to converge within the given tolerance, for each hyperparameter value. If many of these are exactly equal to max_iter, then their solutions have not converged with the tolerance specified by obj_tol. In which case, it is often prudent to increase max_iter.

training_errors

Mean-squared error between the observed outputs y and the trend filtering estimate, for every hyperparameter choice.

optimisms

SURE-estimated optimisms, i.e. optimisms = generalization_errors - training_errors.

x

Vector of observed inputs.

y

Vector of observed outputs.

weights

Weights for the observed outputs, defined as the reciprocal variance of the additive noise that contaminates the signal.

fitted_values

Optimized trend filtering estimate, evaluated at the observed inputs x.

residuals

residuals = y - fitted_values

k

Degree of the trend filtering estimator.

admm_params

List of parameter settings for the trend filtering ADMM algorithm, constructed by passing the optimization_params list to glmgen::trendfilter.control.list().

thinning

Logical. If TRUE, then the data are preprocessed so that a smaller, better conditioned data set is used for fitting.

x_scale, y_scale, data_scaled

For internal use.

Details

Our recommendations for when to use cv_trendfilter() vs. sure_trendfilter() are shown in the table below.

A regularly-sampled data set with some discarded pixels (either sporadically or in large consecutive chunks) is still considered regularly sampled. When the inputs are regularly sampled on a transformed scale, we recommend transforming to that scale and carrying out the full trend filtering analysis on that scale. See the example below for a case when the inputs are evenly sampled on the log10(x) scale.

Scenario	Hyperparameter optimization

x is irregularly sampled	cv_trendfilter()

x is regularly sampled and reciprocal variances are not available	cv_trendfilter()

x is regularly sampled and reciprocal variances are available	sure_trendfilter()

Trend filtering with Stein's unbiased risk estimate

Here we describe the general motivation for optimizing a trend filtering estimator with respect to Stein's unbiased risk estimate. See Politsch et al. (2020a) for more details.

Suppose we observe noisy measurements of an output variable of interest (e.g., flux, magnitude, photon counts) according to the data generating process \[y_i = f(x_i) + \epsilon_i, \quad\quad x_1,\dots,x_n\in(a,b),\] where \(y_i\) is a noisy observation of a signal \(f(x_i)\) and the \(\epsilon_i\) have mean zero with variance \(\sigma_{i}^{2} = \text{Var}(\epsilon_{i})\). Let \(\hat{f}(\cdot\;; \lambda)\) denote the trend filtering estimator of order \(k\) with tuneable hyperparameter \(\lambda\). The fixed-input mean-squared prediction error (MSPE) of the estimator \(\hat{f}\) is defined as \[R(\lambda) = \frac{1}{n}\sum_{i=1}^{n}\;\mathbb{E}\left[\left(y_i - \hat{f}(x_{i};\lambda)\right)^2\;|\;x_{1},\dots,x_{n}\right]\] \[= \frac{1}{n}\sum_{i=1}^{n}\left(\mathbb{E}\left[\left(f(x_i) - \hat{f}(x_i;\lambda)\right)^2\;|\;x_1,\dots,x_n\right] + \sigma_i^2\right).\] Stein's unbiased risk estimate (SURE) provides an unbiased estimate of the fixed-input MSPE via the following formula: \[\hat{R}(\lambda) = \frac{1}{n}\sum_{i=1}^{n}\big(y_i - \hat{f}(x_i; \lambda)\big)^2 + \frac{2\overline{\sigma}^{2}\text{df}(\hat{f})}{n},\] where \(\overline{\sigma}^{2} = n^{-1}\sum_{i=1}^{n} \sigma_i^2\) and \(\text{df}(\hat{f})\) is the effective degrees of freedom of the trend filtering estimator (with a fixed choice of hyperparameter). The generalized lasso results of Tibshirani and Taylor (2012) provide the following formula for the effective degrees of freedom of a trend filtering estimator (with a fixed hyperparameter choice): \[\text{df}(\hat{f}) = \mathbb{E}\left[\text{number of knots in}\;\hat{f}\right] + k + 1.\] The optimal hyperparameter value is then defined as \[\hat{\lambda} = \arg\min_{\lambda} \frac{1}{n}\sum_{i=1}^{n}\big(y_i - \hat{f}(x_i; \lambda)\big)^2 + \frac{2\hat{\overline{\sigma}}^{2}\hat{\text{df}}(\hat{f})}{n},\] where \(\hat{\text{df}}\) is the estimate for the effective degrees of freedom that is obtained by replacing the expectation with the observed number of knots, and \(\hat{\overline{\sigma}}^2\) is an estimate of \(\overline{\sigma}^2\). We define \(\overline{\sigma}^2\) as mean(1 / weights), so weights must be passed in order to use sure_trendfilter. If a reliable estimate of \(\overline{\sigma}^2\) is not available a priori, a data-driven estimate can be constructed, e.g. see Wasserman (2004) or Hastie, Tibshirani, and Friedman (2009).

References

Companion references

Politsch et al. (2020a). Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy. MNRAS, 492(3), p. 4005-4018.

Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical signals with varying degrees of smoothness. MNRAS, 492(3), p. 4019-4032.

Stein's unbiased risk estimate

Tibshirani and Wasserman (2015). Stein’s Unbiased Risk Estimate. 36-702: Statistical Machine Learning course notes (Carnegie Mellon University).

Efron (2014). The Estimation of Prediction Error: Covariance Penalties and Cross-Validation. Journal of the American Statistical Association. 99(467), p. 619-632.

Stein (1981). Estimation of the Mean of a Multivariate Normal Distribution. The Annals of Statistics. 9(6), p. 1135-1151.

Effective degrees of freedom for trend filtering

Tibshirani and Taylor (2012). Degrees of freedom in lasso problems. The Annals of Statistics, 40(2), p. 1198-1232.

See also

Examples

data(quasar_spectrum)
head(spec)

#> # A tibble: 6 × 3
#>   log10_wavelength   flux weights
#>              <dbl>  <dbl>   <dbl>
#> 1             3.55  0.424  0.0417
#> 2             3.55 -2.11   0.125 
#> 3             3.55 -3.78   0.128 
#> 4             3.55  3.60   0.121 
#> 5             3.55 -4.89   0.134 
#> 6             3.55  1.76   0.132 

if (FALSE) {
sure_tf <- sure_trendfilter(spec$log10_wavelength, spec$flux, spec$weights)
}